마이페이지 기능구현. 가입시 쿠폰, 포인트 지급 기능

+++ src/main/java/com/sh/vo/User.java
@@ -33,7 +33,7 @@ public class User {
 		private Date updatedDate;
 		private String deleted;
 		private Date deletedDate;
-		private double point;
+		private int point;
   	    private int stay;
 	    private Grade grade;
 	    private Date birthDay;
